chore(with-storage-sync): move with-storage-sync(and test) to with-storage-sync folders
